requirements.txt
pandas==2.2.2
requests==2.31.0
tqdm==4.66.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
numpy==1.26.2
fastapi==0.115.4
uvicorn==0.31.0
